fix api unit tests
